+++ docs/source/changelog.rst
@@ -36,6 +36,7 @@ Changed
 - (Breaking) Drop support of `deepinv.train` in favor of :class:`deepinv.Trainer` (:gh:`1210` by `Minh Hai Nguyen`_)
 - (Breaking) Drop support of the deprecated parameter `eval_n_samples` in :meth:`deepinv.loss.SplittingLoss.adapt_model` (:gh:`1210` by `Minh Hai Nguyen`_)
 - (Breaking) Drop support of the deprecated parameter `fast` :class:`deepinv.physics.CompressedSensing` (:gh:`1210` by `Minh Hai Nguyen`_)
+- (Breaking) Drop support of `deepinv.Trainer.log_metrics_wandb` (:gh:`1210` by `Minh Hai Nguyen`_)
 
 Fixed
 ^^^^^